SOF-400: Implement functionality for model inference toggle inside data collection session
