--- app/src/main/java/com/vci/vectorcamapp/imaging/domain/model/LiveFrameProcessingResult.kt
@@ -1,10 +1,0 @@
-package com.vci.vectorcamapp.imaging.domain.model
-
-import androidx.compose.ui.geometry.Offset
-import com.vci.vectorcamapp.core.domain.model.InferenceResult
-
-data class LiveFrameProcessingResult(
-    val specimenId: String = "",
-    val autofocusPoint: Offset? = null,
-    val previewInferenceResults: List<InferenceResult> = emptyList()
-)
