+++ app/src/main/java/com/vci/vectorcamapp/imaging/domain/strategy/ImagingWorkflowFactory.kt
@@ -6,13 +6,11 @@ import com.vci.vectorcamapp.imaging.domain.strategy.concrete.DataCollectionImagi
 import com.vci.vectorcamapp.imaging.domain.strategy.concrete.SurveillanceImagingWorkflow
 import javax.inject.Inject
 
-class ImagingWorkflowFactory @Inject constructor(
-    private val inferenceRepository: InferenceRepository
-) {
+class ImagingWorkflowFactory @Inject constructor() {
     fun create(sessionType: SessionType): ImagingWorkflow {
         return when (sessionType) {
-            SessionType.SURVEILLANCE -> SurveillanceImagingWorkflow(inferenceRepository)
-            SessionType.DATA_COLLECTION -> DataCollectionImagingWorkflow(inferenceRepository)
+            SessionType.SURVEILLANCE -> SurveillanceImagingWorkflow()
+            SessionType.DATA_COLLECTION -> DataCollectionImagingWorkflow()
         }
     }
 }
