+++ app/src/main/java/com/vci/vectorcamapp/imaging/domain/strategy/concrete/DataCollectionImagingWorkflow.kt
@@ -1,35 +1,11 @@
 package com.vci.vectorcamapp.imaging.domain.strategy.concrete
 
-import android.graphics.Bitmap
-import com.vci.vectorcamapp.core.domain.util.Result
-import com.vci.vectorcamapp.imaging.domain.model.CapturedFrameProcessingResult
-import com.vci.vectorcamapp.imaging.domain.model.LiveFrameProcessingResult
-import com.vci.vectorcamapp.imaging.domain.repository.InferenceRepository
 import com.vci.vectorcamapp.imaging.domain.strategy.ImagingWorkflow
-import com.vci.vectorcamapp.imaging.domain.util.ImagingError
-import javax.inject.Inject
 
-class DataCollectionImagingWorkflow @Inject constructor(
-    private val inferenceRepository: InferenceRepository,
-) : ImagingWorkflow {
+class DataCollectionImagingWorkflow : ImagingWorkflow {
     override val specimenFurtherProcessingProbability: Float
         get() = 0f
 
     override val allowModelInferenceToggle: Boolean
         get() = true
-
-    override suspend fun processLiveFrame(bitmap: Bitmap): LiveFrameProcessingResult {
-        return LiveFrameProcessingResult(
-            specimenId = inferenceRepository.readSpecimenId(bitmap),
-            previewInferenceResults = emptyList()
-        )
-    }
-
-    override suspend fun processCapturedFrame(bitmap: Bitmap): Result<CapturedFrameProcessingResult, ImagingError> {
-        return Result.Success(CapturedFrameProcessingResult())
-    }
-
-    override fun close() {
-        inferenceRepository.closeResources()
-    }
 }
